refactored test_resources.py into test_resources package
